moved test and updated end point
